Html, CSS and JavaScript File
This project is a basic calculator featuring addition, subtraction, multiplication, and division, designed to aid beginners with essential HTML, CSS, and JavaScript skills. The user interface includes buttons for data entry and a display screen for results. It demonstrates the practical application of web development skills in creating a functional, user-friendly tool.